feat: add initial web platform support with stub implementation
Extract ChannelClient into client_io.dart and use conditional imports
to select the platform client at compile time. Add client_web.dart
with stub implementations that print method names for start(), setUser(),
and notify(). Register the web platform in pubspec.yaml with a no-op
plugin entry point.
