+++ packages/bugsnag_flutter/lib/bugsnag_flutter_web.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+
+class BugsnagFlutterWeb {
+  static void registerWith(Registrar registrar) {
+    // No-op: web implementation uses dart:js_interop, not MethodChannel
+  }
+}
