+++ packages/bugsnag_flutter/test/channel_client_test.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:bugsnag_flutter/bugsnag_flutter.dart';
-import 'package:bugsnag_flutter/src/client.dart';
+import 'package:bugsnag_flutter/src/client_io.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'test_channel.dart';
